This update gives the installer some love: Made the installer use smarty - this makes changing the installer visuals in the future much easier, should somebody feel frisky. The exception to this is most of step 4, as the logic and the UI are tightly bound together and are difficult to separate. I'll prod this in the future if I can work out WTH is going on in there. The installer templates are in /install/templates & the compiled templates in /install/templates_c.
Fixed up the grammar a fair amount in the installer, and changed some of the descriptions.
Fixed the layout a bit in the installer, and added a few visual enhancements - stuff should be easier to see.
Installation errors are a tad more visual.
Re-activated / deleted dead-code where it made sense.
Added a test for cURL.

This could be a tad unstable, there's only so many tests I can run :) Feedback is welcome.