Adjusted step 1 of the installer to include projects that EDK makes use of, eg Smarty. Added the auto-updater.
Here's a short summary on how it works:
Login as an admin.
As you login, you will be presented by a couple of links depending on what kind of update we are providing, but essentially they'll redirect you to the admin_upgrade page.
You will then be presented all of the existing upgrades that you can download and apply. Database updates will come in the form of simple SQL scripts mainly for adding extra values to a table like kb3_invTypes, as an example.
For more advanced database upgrades, Kovell's existing database upgrade engine can be used, as it is robust and tolerant of many operations that need to occur. Altering table structures or data will still be done in this manner - all you need is to add the relevant files into a code upgrade zip, explained next.
Code upgrades come in the form of a zip that is packed to the exact directory structure of the killboard - upon the application of a code upgrade, the updater will backup all of the files that will be overwritten (so that you can manually roll back if need be), and apply whatever code changes will occur, and will automatically purge the template cache, and the currently cached queries.

File locations are bound to change. Feedback definetly welcome. I'll also add a more detailed explanation of how it all works on the google edk project's wiki.