Changed the way aggressive structures are handled on killmails - The name of the corporation who owns the tower / mobile warp disruptor is appended to the front of the relevant structure. (Fixes issue 76 - bubbles and towers should no longer swap allegiances every fight.)
Updated some of the awards to use the better looking variants and in some cases entirely replacing the old ones. (That old eagle award was really starting to look crappy.)

Searched through the entire source and culled images that are no longer being used anywhere. Still to go would be the old panel graphics and possibly running through the CSS files to remove unused entries.