Check for needed upgrade of CCP DB after KB upgrade
The CCP DB is now included in the KB releases. The CCP DB version number
is defined in globals.h upon release.
If a new release is installed with a new CCP DB, this will now be detected
and and URL to reinstall/update the CCP DB is provided.
Upgrading CCP DB uses the already available CCP DB reset mechanism of the board.