Enhancement: Check file permissions before update
The auto updater now checks for write permissions for every file affected
by the update before touching anything. If file permissions are not
sufficient, the updater tries to change them. If that didn't work it fails
with a list of all files which need to be set writeable by the user in
order to complete the update.
The update will not be applied partially anymore.