Fix: Fatal error on killmail import
Fixed fatal error on killmail import (IDFeed/API) due to unknown typeIDs
when creating the raw mail for hashing a mail. This behavior used to block
the import completely.
The mail with error will now be skipped, an error message with details for
the nature of the error will be displayed.