Fix: Add corp name in front of control towers/mods
Fixed a bug in IDFeed/API import that identified Control Towers as
involved parties as NPCs and therefore ignored kills with towers as the
only involved party if configured to ignore NPC only kills.

Side effect of this fix is that the pilots representing the towers in the
database are now called <corp> - <tower> once again.