Fix: Adding corps/alliances as owners by external ID
Same fix as for pilots: Fixed the check for correct ID range for external
ID of corporations and alliances when adding a new killboard owner by
external IDs.
Corporation IDs have to be above 98 million, alliance IDs above 99 million
since the 64bit change (see
http://forums.eveonline.com/default.aspx?g=posts&m=716708#post716708).
Also fixed adding new owners by name for names starting with numbers. This
was regarded as ID and unsuccessful.

Thanks for the fix to Bubalubs.