Enhancement: Negative timestamp offset for zKB fetch
Removed the setting for how many kills to fetch from zKB, since the zKB
API no longer supports that modifier.

Added a negative offset for the last kill timestamp. This offset will be
applied to the last kill timestamp for each fetch in order to fetch kills
that were posted after the latest kill fetched in the previous cycle.