Fix: zKB sliding timestamp and HTTP errors
When an error occurred while fetching from zKB the last kill timestamp for
the zKB fetch configuration was modified by the value of the negative
timestamp offset, sliding more and more into the past with each error.

Moreover, an HTTP 500 error on fetch was not regarded as one and an empty
data set was returned. With this commit, HTTP 500 return codes are
regarded as errors.