IDFeed 1.3 / EDK 4.2.8.0
Updated IDFeed to version 1.3 by removing the trust functionality.
If a kill comes with with an external ID via IDFeed, it is consideres API
verified.
Kept the trusted database field in kb3_mails for "delete permanently"
functionality, which uses this field with a value of -1.

Updated constants.php for EDK 4.2.8.0