Fix for infamous "Invalid argument supplied..."
Fix for the infamous warnings message "Invalid argument supplied for
foreach()" in alliance detail view and API KillLog in ACP.
This happened when the API returned an error or the connection to the API
could not be established for whatever reason.

Sanity-checks have been added so the data is checked first before working
on it.
Problems with the API connection can be seen in the troubleshooting
section in ACP.