Enhancement: Added support for item singleton flags
Added proper support for singleton flags of items on kills.
This flag is mainly used to indicate whether an item was a blueprint copy
or original.
In the past this information has been stored as special location flag,
indicating a blueprint copy which would always be placed inside the
cargohold.
If a BPC was located inside a special bay (e.g. fleet hangar), it would
show up in the cargo on the killboard.