Rewrote table updates for better compatibility
Rewrote the table updates to kb3_items_dropped and kb3_items_destroyed to
be executed in chunks.
The new table structure will be created as temporary table. Then all
entries from the old table are copied to the temporary table. In the end,
the old table gets deleted and the temporary table renamed.

Adding a column to tables with so many entries might exceed the script
execution time on some servers, so the update gets done in chunks.