Updated zKB fetch example URLs
to include the "combined" modifier. This new API route had been introduced
into the zKillboard branch of eve-kill first (beta.eve-kill.net) and is
mandatory for the version running there, zKillboard.com applied it as
optional. Changed the example text to include the "combined" route to make
the example valid for both zKillboard.com and beta.eve-kill.net