Update class.translate.php
Fixes this issue:
http://www.evekb.org/forum/viewtopic.php?f=503&t=18821&p=42639&hilit=Mechanicsal#p42639

Apparently the preparser translation changes "Mechanical Parts" to "Mechanicsal Parts" for pre-cru1.0 killmails. The actual typeName change in this relase was "Mechanic" to "Mechanics" instead. Moving that replacement from str_replace to preg_replace appears to resolve it (by checking that "Mechanic" is not followed by "a"). There is likely a more elegant regex usable for this but I couldn't think of one right now.