Feature: Added kill location within systems
The kill detail page will now display the nearest celestial of the kill,
including the distance from it (m, km, k km, M km, AU). This will only be
displayed if the coordinates are available for the kill (post Parallax).

Also added kill coordinates to the IDFeed and increased IDFeed version to
1.4.