Fix: CREST link calculations for Drifter kills
Fixed calculation of the CREST links for kills with Drifters as final blow
pilots.
The calculation requires the external Pilot ID of the final blow pilot.
EDK contains a sanity check for character IDs (must be > 90000000).
Drifters have Character-IDs in the range of 3000000. The sanity check has
been modified accordingly.