Enhancement: Fetch all kills via CREST
Information obtained via API, IDFeed and zKB Fetch is used to construct
the kill's CREST link. The kill itself is then fetched via CREST to
guarantee data accuracy. If an invalid CREST link is constructed, there's
an option to still post the kill without API verification.

Since the CREST response time is 500ms on average for a random kill, the
maximum number of kills to fetch from CREST can be limited in the ACP.
This value is automatically determined on first use by evaluating the PHP
time limit.