Fix: Fetching item types via CREST
If item types on kills are not found in the database, EDK tries to pull
them via the CREST types endpoint. However, it only accepted items that
got a description.
Recently new SKINs turned up that were not included in any previous SDEs
and did not have a description set, so EDK did not accept them.
The check for description has been removed.