Enhancement: Try to use CREST hash provided by zKB first before parsing kills
ZKBFetch now directly uses the CREST hash and external ID provided by zKB
to construct the CREST URL of a kill, instead of parsing the kill data and
calculating the CREST has by itself.
In addition, kills get skipped with a proper error message if CREST
returns an error for a kill and "skip non-verifiable kills" option is
enabled.