Fix: ZKBFetch now accepts formats generated by XML -> CREST converter
In order to work around an error thrown by CREST if a kill contains an
entity with a faction, zKB converts XML API kills to CREST and parses
these kills (if available).
The output of that conversions differs in some cases from what the CREST
API generates (if a weapon/corporation is missing for an involved party in
CREST output, zKB returns an ID of 0).