Fix for zKB Fetch for entities with > 200 kills/h
Fixed a problem caused by the zKB API changes, that caused zKB Fetch to
not advance the last kill timestamp if the last kill in the fetched
batch did not occur in the hour following the start timestamp.

Introduced using the page modifier to keep fetching until either no
more kills are returned or a kill in the following hour is found.