Fix: Entity summary table stats after deleting kills
If a kill with the victim and involved parties belonging to the same
entity was deleted, the kill count as well as the loss count for that
ship class was decremented in the summary table for that entity (even
though the summary table stats were computed correctly before).

Such kills are treated as losses in the summary tables for individual
entities, deleting them now only decrements the loss count for that ship
class.