Enhancement: Added keep-alive like behavior to ESI library
The EDK ESI library now uses curl_multi_init() (if available) in order
to re-use the connection to the ESI servers.
Since a multitide of parameters could be set, re-using a simple cURL
handle is not practical.
This optimization reduces the time spent talking to ESI for fetching 60
kills (and resolving all entity IDs) by over 25%.