ZKBFetch: Fixed bug leading to potential hang while fetching
Fixed a bug in zKBFetch that could cause a specific fetch configuration
to not advance when fetching, but to get stuck.
If a kill turns out to be a duplicate while posting, the kill was still
counted as "posted". If the maximum number of kills is reached with such
kills, and all previously processed kills have a timestamp within the
same minute, the last kill timestamp will not move.

This fix only counts kills as posted if they were actually new kills.