ESI: Standardized HTTP response headers with lower case
HTTP headers are case insensitive, so when using them, we first need to
standardize the case before trying and accessing them.

Might also fix a problem causing ESI Fetch to only fetch the first page
of kills in some cases (because the number of pages is read from the HTTP response
headers).