ZKBFetch: Fix for fetching kill history from zKillboard
When the order modifier of zKB API was disabled, killsmails would be
returned in descending order time. That was incompatible with EDK's
fetching algorithm.
The fix implemented back then was not sufficient for working with last
killtimestamps in the past.
The current fix employs the use of the endTime modifier, if the
starTime modifier is more than one day in the past. the endTime modifier
is then set to 1 day nearer to the future. Kills are fetched one day at
a time. Each cycle advances the kill timestamp by a single day at most.