Fix: Post kills within the classification time frame
If kill classification was enabled, a bug prevented kills from being
posted if the kill timestamp was still within the classification window.

The posting routing would try to calculate the hash from the old text
mail format. However, when calling the getRawMail() method in the kill
class, an error string would be returned if the kill was still
classified. The parser then was unable to create a hash, and without a
hash value, the kill could not be posted.
The fix checks if the kill is actually posted when checking if it is
classified. Kills that are not yet posted, cannot be classified.